requirements.txt
flask==1.1.2
authlib==0.15.3
Flask-Discord==0.1.64
gunicorn==20.0.4
werkzeug==1.0.1
discord-webhook==0.8.0
raygun4py==4.3.0
gql~=3.0.0a5
PyJWT~=2.0.1
